Chore: reorganized members of the editor
