+++ lib/src/editor/raw_editor/boundaries/editor_text_boundaries.dart
@@ -1,0 +1,12 @@
+// combinators
+export 'combinators/quill_editor_collapsed_selection_boundary.dart';
+export 'combinators/quill_editor_mixed_boundary.dart';
+export 'combinators/quill_editor_text_expanded_boundary.dart';
+// base
+export 'quill_editor_text_boundary.dart';
+// text
+export 'text/quill_editor_character_boundary.dart';
+export 'text/quill_editor_document_boundary.dart';
+export 'text/quill_editor_line_break_boundary.dart';
+export 'text/quill_editor_whitespace_boundary.dart';
+export 'text/quill_editor_word_boundary.dart';
