+++ lib/src/editor/raw_editor/config/raw_editor_config.dart
@@ -7,11 +7,11 @@ import '../../../document/nodes/node.dart';
 import '../../../editor/embed/embed_editor_builder.dart';
 import '../../../editor/raw_editor/raw_editor.dart';
 import '../../../editor/raw_editor/raw_editor_state.dart';
-import '../../../editor/widgets/cursor.dart';
-import '../../../editor/widgets/default_styles.dart';
 import '../../../editor/widgets/delegate.dart';
 import '../../../editor/widgets/link.dart';
+import '../../../editor/widgets/styles/default_styles.dart';
 import '../../../toolbar/theme/quill_dialog_theme.dart';
+import '../../widgets/styles/cursor_style.dart';
 import '../../widgets/text/utils/text_block_utils.dart';
 import '../builders/leading_block_builder.dart';
 import 'events/events.dart';
