+++ lib/src/editor/raw_editor/editor_glyph_heights.dart
@@ -1,0 +1,12 @@
+import 'package:meta/meta.dart';
+
+@immutable
+class QuillEditorGlyphHeights {
+  const QuillEditorGlyphHeights(
+    this.startGlyphHeight,
+    this.endGlyphHeight,
+  );
+
+  final double startGlyphHeight;
+  final double endGlyphHeight;
+}
