+++ lib/src/editor/raw_editor/editor_state.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/widgets.dart';
+import '../../common/structs/offset_value.dart';
+import '../render/render_editor.dart';
+import '../widgets/text/selection/text_selection.dart';
+import 'raw_editor.dart';
+
+/// Base interface for the editor state which defines contract used by
+/// various mixins.
+abstract class EditorState extends State<QuillRawEditor>
+    implements TextSelectionDelegate {
+  ScrollController get scrollController;
+
+  RenderEditor get renderEditor;
+
+  EditorTextSelectionOverlay? get selectionOverlay;
+
+  List<OffsetValue> get pasteStyleAndEmbed;
+
+  String get pastePlainText;
+
+  /// Controls the floating cursor animation when it is released.
+  /// The floating cursor is animated to merge with the regular cursor.
+  AnimationController get floatingCursorResetController;
+
+  /// Returns true if the editor has been marked as needing to be rebuilt.
+  bool get dirty;
+
+  bool showToolbar();
+
+  void requestKeyboard();
+}
