+++ lib/src/editor/raw_editor/keyboard_shortcuts/editor_keyboard_shortcut_actions.dart
@@ -4,10 +4,10 @@ import '../../../document/attribute.dart';
 import '../../../document/style.dart';
 import '../../../toolbar/buttons/link_style2_button.dart';
 import '../../../toolbar/buttons/search/search_dialog.dart';
-import '../../editor.dart';
+import '../../render/utils/quill_vertical_caret_movement.dart';
 import '../../widgets/link.dart';
+import '../boundaries/editor_text_boundaries.dart';
 import '../raw_editor_state.dart';
-import '../raw_editor_text_boundaries.dart';
 
 // -------------------------------  Text Actions -------------------------------
 class QuillEditorDeleteTextAction<T extends DirectionalTextEditingIntent>
