+++ lib/src/editor/raw_editor/raw_editor.dart
@@ -1,9 +1,6 @@
 import 'package:flutter/widgets.dart';
 
-import '../../common/structs/offset_value.dart';
 import '../../controller/quill_controller.dart';
-import '../editor.dart';
-import '../widgets/text/text_selection.dart';
 import 'config/raw_editor_config.dart';
 import 'raw_editor_state.dart';
 
@@ -40,40 +37,3 @@ typedef QuillEditorContextMenuBuilder = Widget Function(
   BuildContext context,
   QuillRawEditorState rawEditorState,
 );
-
-@immutable
-class QuillEditorGlyphHeights {
-  const QuillEditorGlyphHeights(
-    this.startGlyphHeight,
-    this.endGlyphHeight,
-  );
-
-  final double startGlyphHeight;
-  final double endGlyphHeight;
-}
-
-/// Base interface for the editor state which defines contract used by
-/// various mixins.
-abstract class EditorState extends State<QuillRawEditor>
-    implements TextSelectionDelegate {
-  ScrollController get scrollController;
-
-  RenderEditor get renderEditor;
-
-  EditorTextSelectionOverlay? get selectionOverlay;
-
-  List<OffsetValue> get pasteStyleAndEmbed;
-
-  String get pastePlainText;
-
-  /// Controls the floating cursor animation when it is released.
-  /// The floating cursor is animated to merge with the regular cursor.
-  AnimationController get floatingCursorResetController;
-
-  /// Returns true if the editor has been marked as needing to be rebuilt.
-  bool get dirty;
-
-  bool showToolbar();
-
-  void requestKeyboard();
-}
