+++ lib/src/editor/raw_editor/render/raw_editor_render_object.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart' show ViewportOffset;
 
-import '../../document/document.dart';
-import '../editor.dart';
-import '../widgets/cursor.dart';
+import '../../../document/document.dart';
+import '../../render/render_editor.dart';
+import '../../render/utils/render_editor_utils.dart';
+import '../../widgets/cursor.dart';
 
 class QuillRawEditorMultiChildRenderObject
     extends MultiChildRenderObjectWidget {
