+++ lib/src/editor/render/render_editable_container_box.dart
@@ -1,0 +1,239 @@
+import 'dart:math' as math;
+import 'package:flutter/rendering.dart';
+import 'package:meta/meta.dart';
+
+import '../../document/nodes/container.dart' as container_node;
+import '../widgets/box.dart';
+
+class EditableContainerParentData
+    extends ContainerBoxParentData<RenderEditableBox> {}
+
+/// Multi-child render box of editable content.
+///
+/// Common ancestor for [RenderEditor] and [RenderEditableTextBlock].
+class RenderEditableContainerBox extends RenderBox
+    with
+        ContainerRenderObjectMixin<RenderEditableBox,
+            EditableContainerParentData>,
+        RenderBoxContainerDefaultsMixin<RenderEditableBox,
+            EditableContainerParentData> {
+  RenderEditableContainerBox({
+    required this.container,
+    required this.textDirection,
+    required this.scrollBottomInset,
+    required EdgeInsetsGeometry padding,
+    List<RenderEditableBox>? children,
+  })  : assert(padding.isNonNegative),
+        _padding = padding {
+    addAll(children);
+  }
+
+  @protected
+  container_node.QuillContainer container;
+  TextDirection textDirection;
+  EdgeInsetsGeometry _padding;
+  double scrollBottomInset;
+  EdgeInsets? _resolvedPadding;
+
+  void setContainer(container_node.QuillContainer c) {
+    if (container == c) {
+      return;
+    }
+    container = c;
+    markNeedsLayout();
+  }
+
+  EdgeInsetsGeometry getPadding() => _padding;
+
+  void setPadding(EdgeInsetsGeometry value) {
+    assert(value.isNonNegative);
+    if (_padding == value) {
+      return;
+    }
+    _padding = value;
+    _markNeedsPaddingResolution();
+  }
+
+  EdgeInsets? get resolvedPadding => _resolvedPadding;
+
+  void resolvePadding() {
+    if (_resolvedPadding != null) {
+      return;
+    }
+    _resolvedPadding = _padding.resolve(textDirection);
+    _resolvedPadding = _resolvedPadding!.copyWith(left: _resolvedPadding!.left);
+
+    assert(_resolvedPadding!.isNonNegative);
+  }
+
+  RenderEditableBox childAtPosition(TextPosition position) {
+    assert(firstChild != null);
+    final targetNode = container.queryChild(position.offset, false).node;
+
+    var targetChild = firstChild;
+    while (targetChild != null) {
+      if (targetChild.container == targetNode) {
+        break;
+      }
+      final newChild = childAfter(targetChild);
+      if (newChild == null) {
+        //  At start of document fails to find the position
+        targetChild = childAtOffset(const Offset(0, 0));
+        break;
+      }
+      targetChild = newChild;
+    }
+    if (targetChild == null) {
+      throw 'targetChild should not be null';
+    }
+    return targetChild;
+  }
+
+  void _markNeedsPaddingResolution() {
+    _resolvedPadding = null;
+    markNeedsLayout();
+  }
+
+  /// Returns child of this container located at the specified local `offset`.
+  ///
+  /// If `offset` is above this container (offset.dy is negative) returns
+  /// the first child. Likewise, if `offset` is below this container then
+  /// returns the last child.
+  RenderEditableBox childAtOffset(Offset offset) {
+    assert(firstChild != null);
+    resolvePadding();
+
+    if (offset.dy <= _resolvedPadding!.top) {
+      return firstChild!;
+    }
+    if (offset.dy >= size.height - _resolvedPadding!.bottom) {
+      return lastChild!;
+    }
+
+    var child = firstChild;
+    final dx = -offset.dx;
+    var dy = _resolvedPadding!.top;
+    while (child != null) {
+      if (child.size.contains(offset.translate(dx, -dy))) {
+        return child;
+      }
+      dy += child.size.height;
+      child = childAfter(child);
+    }
+
+    // this case possible, when editor not scrollable,
+    // but minHeight > content height and tap was under content
+    return lastChild!;
+  }
+
+  @override
+  void setupParentData(RenderBox child) {
+    if (child.parentData is EditableContainerParentData) {
+      return;
+    }
+
+    child.parentData = EditableContainerParentData();
+  }
+
+  @override
+  void performLayout() {
+    assert(constraints.hasBoundedWidth);
+    resolvePadding();
+    assert(_resolvedPadding != null);
+
+    var mainAxisExtent = _resolvedPadding!.top;
+    var child = firstChild;
+    final innerConstraints =
+        BoxConstraints.tightFor(width: constraints.maxWidth)
+            .deflate(_resolvedPadding!);
+    while (child != null) {
+      child.layout(innerConstraints, parentUsesSize: true);
+      final childParentData = (child.parentData as EditableContainerParentData)
+        ..offset = Offset(_resolvedPadding!.left, mainAxisExtent);
+      mainAxisExtent += child.size.height;
+      assert(child.parentData == childParentData);
+      child = childParentData.nextSibling;
+    }
+    mainAxisExtent += _resolvedPadding!.bottom;
+    size = constraints.constrain(Size(constraints.maxWidth, mainAxisExtent));
+
+    assert(size.isFinite);
+  }
+
+  double _getIntrinsicCrossAxis(double Function(RenderBox child) childSize) {
+    var extent = 0.0;
+    var child = firstChild;
+    while (child != null) {
+      extent = math.max(extent, childSize(child));
+      final childParentData = child.parentData as EditableContainerParentData;
+      child = childParentData.nextSibling;
+    }
+    return extent;
+  }
+
+  double _getIntrinsicMainAxis(double Function(RenderBox child) childSize) {
+    var extent = 0.0;
+    var child = firstChild;
+    while (child != null) {
+      extent += childSize(child);
+      final childParentData = child.parentData as EditableContainerParentData;
+      child = childParentData.nextSibling;
+    }
+    return extent;
+  }
+
+  @override
+  double computeMinIntrinsicWidth(double height) {
+    resolvePadding();
+    return _getIntrinsicCrossAxis((child) {
+      final childHeight = math.max<double>(
+          0, height - _resolvedPadding!.top + _resolvedPadding!.bottom);
+      return child.getMinIntrinsicWidth(childHeight) +
+          _resolvedPadding!.left +
+          _resolvedPadding!.right;
+    });
+  }
+
+  @override
+  double computeMaxIntrinsicWidth(double height) {
+    resolvePadding();
+    return _getIntrinsicCrossAxis((child) {
+      final childHeight = math.max<double>(
+          0, height - _resolvedPadding!.top + _resolvedPadding!.bottom);
+      return child.getMaxIntrinsicWidth(childHeight) +
+          _resolvedPadding!.left +
+          _resolvedPadding!.right;
+    });
+  }
+
+  @override
+  double computeMinIntrinsicHeight(double width) {
+    resolvePadding();
+    return _getIntrinsicMainAxis((child) {
+      final childWidth = math.max<double>(
+          0, width - _resolvedPadding!.left + _resolvedPadding!.right);
+      return child.getMinIntrinsicHeight(childWidth) +
+          _resolvedPadding!.top +
+          _resolvedPadding!.bottom;
+    });
+  }
+
+  @override
+  double computeMaxIntrinsicHeight(double width) {
+    resolvePadding();
+    return _getIntrinsicMainAxis((child) {
+      final childWidth = math.max<double>(
+          0, width - _resolvedPadding!.left + _resolvedPadding!.right);
+      return child.getMaxIntrinsicHeight(childWidth) +
+          _resolvedPadding!.top +
+          _resolvedPadding!.bottom;
+    });
+  }
+
+  @override
+  double computeDistanceToActualBaseline(TextBaseline baseline) {
+    resolvePadding();
+    return defaultComputeDistanceToFirstActualBaseline(baseline)! +
+        _resolvedPadding!.top;
+  }
+}
