+++ lib/src/editor/widgets/proxies/baseline_proxy.dart
@@ -1,0 +1,80 @@
+import 'package:flutter/rendering.dart';
+import 'package:flutter/widgets.dart';
+
+class BaselineProxy extends SingleChildRenderObjectWidget {
+  const BaselineProxy({
+    super.key,
+    super.child,
+    this.textStyle,
+    this.padding,
+  });
+
+  final TextStyle? textStyle;
+  final EdgeInsets? padding;
+
+  @override
+  RenderBaselineProxy createRenderObject(BuildContext context) {
+    return RenderBaselineProxy(
+      null,
+      textStyle!,
+      padding,
+    );
+  }
+
+  @override
+  void updateRenderObject(
+      BuildContext context, covariant RenderBaselineProxy renderObject) {
+    renderObject
+      ..textStyle = textStyle!
+      ..padding = padding!;
+  }
+}
+
+class RenderBaselineProxy extends RenderProxyBox {
+  RenderBaselineProxy(
+    RenderParagraph? super.child,
+    TextStyle textStyle,
+    EdgeInsets? padding,
+  ) : _prototypePainter = TextPainter(
+            text: TextSpan(text: ' ', style: textStyle),
+            textDirection: TextDirection.ltr,
+            strutStyle:
+                StrutStyle.fromTextStyle(textStyle, forceStrutHeight: true));
+
+  final TextPainter _prototypePainter;
+
+  set textStyle(TextStyle value) {
+    if (_prototypePainter.text!.style == value) {
+      return;
+    }
+    _prototypePainter.text = TextSpan(text: ' ', style: value);
+    markNeedsLayout();
+  }
+
+  EdgeInsets? _padding;
+
+  set padding(EdgeInsets value) {
+    if (_padding == value) {
+      return;
+    }
+    _padding = value;
+    markNeedsLayout();
+  }
+
+  @override
+  double computeDistanceToActualBaseline(TextBaseline baseline) =>
+      _prototypePainter.computeDistanceToActualBaseline(baseline);
+  // SEE What happens + _padding?.top;
+
+  @override
+  void performLayout() {
+    super.performLayout();
+    _prototypePainter.layout();
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    _prototypePainter.dispose();
+  }
+}
