+++ lib/src/editor/widgets/proxies/embed_proxy.dart
@@ -1,0 +1,53 @@
+import 'package:flutter/rendering.dart';
+import 'package:flutter/widgets.dart';
+import '../box.dart';
+
+class EmbedProxy extends SingleChildRenderObjectWidget {
+  const EmbedProxy(Widget child, {super.key}) : super(child: child);
+
+  @override
+  RenderEmbedProxy createRenderObject(BuildContext context) =>
+      RenderEmbedProxy(null);
+}
+
+class RenderEmbedProxy extends RenderProxyBox implements RenderContentProxyBox {
+  RenderEmbedProxy(super.child);
+
+  @override
+  List<TextBox> getBoxesForSelection(TextSelection selection) {
+    if (!selection.isCollapsed) {
+      return <TextBox>[
+        TextBox.fromLTRBD(0, 0, size.width, size.height, TextDirection.ltr)
+      ];
+    }
+
+    final left = selection.extentOffset == 0 ? 0.0 : size.width;
+    final right = selection.extentOffset == 0 ? 0.0 : size.width;
+    return <TextBox>[
+      TextBox.fromLTRBD(left, 0, right, size.height, TextDirection.ltr)
+    ];
+  }
+
+  @override
+  double getFullHeightForCaret(TextPosition position) => size.height;
+
+  @override
+  Offset getOffsetForCaret(TextPosition position, Rect caretPrototype) {
+    assert(
+        position.offset == 1 || position.offset == 0 || position.offset == -1);
+    return position.offset <= 0
+        ? Offset.zero
+        : Offset(size.width - caretPrototype.width, 0);
+  }
+
+  @override
+  TextPosition getPositionForOffset(Offset offset) =>
+      TextPosition(offset: offset.dx > size.width / 2 ? 1 : 0);
+
+  @override
+  TextRange getWordBoundary(TextPosition position) =>
+      const TextRange(start: 0, end: 1);
+
+  @override
+  double get preferredLineHeight => size.height;
+}
