+++ lib/src/editor/widgets/proxies/proxy.dart
@@ -1,0 +1,4 @@
+export 'baseline_proxy.dart';
+export 'embed_proxy.dart';
+export 'render_paragraph_proxy.dart';
+export 'rich_text_proxy.dart';
