+++ lib/src/editor/widgets/proxies/render_paragraph_proxy.dart
@@ -1,0 +1,134 @@
+import 'dart:ui';
+
+import 'package:flutter/rendering.dart';
+
+import '../box.dart';
+
+class RenderParagraphProxy extends RenderProxyBox
+    implements RenderContentProxyBox {
+  RenderParagraphProxy(
+    RenderParagraph? super.child,
+    TextStyle textStyle,
+    TextAlign textAlign,
+    TextDirection textDirection,
+    TextScaler textScaler,
+    StrutStyle strutStyle,
+    Locale locale,
+    TextWidthBasis textWidthBasis,
+    TextHeightBehavior? textHeightBehavior,
+  ) : _prototypePainter = TextPainter(
+          text: TextSpan(text: ' ', style: textStyle),
+          textAlign: textAlign,
+          textDirection: textDirection,
+          textScaler: textScaler,
+          strutStyle: strutStyle,
+          locale: locale,
+          textWidthBasis: textWidthBasis,
+          textHeightBehavior: textHeightBehavior,
+        );
+
+  final TextPainter _prototypePainter;
+
+  set textStyle(TextStyle value) {
+    if (_prototypePainter.text!.style == value) {
+      return;
+    }
+    _prototypePainter.text = TextSpan(text: ' ', style: value);
+    markNeedsLayout();
+  }
+
+  set textAlign(TextAlign value) {
+    if (_prototypePainter.textAlign == value) {
+      return;
+    }
+    _prototypePainter.textAlign = value;
+    markNeedsLayout();
+  }
+
+  set textDirection(TextDirection value) {
+    if (_prototypePainter.textDirection == value) {
+      return;
+    }
+    _prototypePainter.textDirection = value;
+    markNeedsLayout();
+  }
+
+  set textScaler(TextScaler value) {
+    if (_prototypePainter.textScaler == value) {
+      return;
+    }
+    _prototypePainter.textScaler = value;
+    markNeedsLayout();
+  }
+
+  set strutStyle(StrutStyle value) {
+    if (_prototypePainter.strutStyle == value) {
+      return;
+    }
+    _prototypePainter.strutStyle = value;
+    markNeedsLayout();
+  }
+
+  set locale(Locale value) {
+    if (_prototypePainter.locale == value) {
+      return;
+    }
+    _prototypePainter.locale = value;
+    markNeedsLayout();
+  }
+
+  set textWidthBasis(TextWidthBasis value) {
+    if (_prototypePainter.textWidthBasis == value) {
+      return;
+    }
+    _prototypePainter.textWidthBasis = value;
+    markNeedsLayout();
+  }
+
+  set textHeightBehavior(TextHeightBehavior? value) {
+    if (_prototypePainter.textHeightBehavior == value) {
+      return;
+    }
+    _prototypePainter.textHeightBehavior = value;
+    markNeedsLayout();
+  }
+
+  @override
+  RenderParagraph? get child => super.child as RenderParagraph?;
+
+  @override
+  double get preferredLineHeight => _prototypePainter.preferredLineHeight;
+
+  @override
+  Offset getOffsetForCaret(TextPosition position, Rect caretPrototype) =>
+      child!.getOffsetForCaret(position, caretPrototype);
+
+  @override
+  TextPosition getPositionForOffset(Offset offset) =>
+      child!.getPositionForOffset(offset);
+
+  @override
+  double? getFullHeightForCaret(TextPosition position) =>
+      child!.getFullHeightForCaret(position);
+
+  @override
+  TextRange getWordBoundary(TextPosition position) =>
+      child!.getWordBoundary(position);
+
+  @override
+  List<TextBox> getBoxesForSelection(TextSelection selection) => child!
+      .getBoxesForSelection(selection, boxHeightStyle: BoxHeightStyle.max);
+
+  @override
+  void performLayout() {
+    super.performLayout();
+    _prototypePainter.layout(
+        minWidth: constraints.minWidth, maxWidth: constraints.maxWidth);
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    _prototypePainter.dispose();
+  }
+}
